requirements.txt
streamlit>=1.30.0
google-generativeai>=0.8.0
yfinance>=0.2.36
pandas>=2.0.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
html5lib>=1.1
PyPDF2>=3.0.0
google-cloud-firestore
google-cloud-bigquery
google-cloud-storage
google-cloud-pubsub
